Promote Jeffrey JIB and IntelliJ Plugin to standalone docs products
Both tools previously lived inside other products' docs (JIB under Hub,
the IntelliJ Plugin under Microscope). Each is now a first-class product
with its own sidebar section (Overview / Configuration / Setup) and a card
on the /docs landing page, mirroring the Provisioner structure.

- Split jib/JeffreyJibPage.vue into JibOverviewPage, JibConfigurationPage
  and JibSetupPage; the Hub 'Deployment by Example' JIB page stays and is
  cross-linked from the new product sidebar.
- Move the IntelliJ plugin pages into views/docs/intellij-plugin/ as
  Overview / Configuration / Setup plus the Java JFR Profiler page; leave a
  cross-link entry in the Microscope sidebar.
- Register both products in useDocsNavigation (types, segments, nav arrays,
  product resolution) and add routes plus back-compat redirects from the
  old URLs.
- Add quinary/senary DocsProductCard variants and two cards on DocsIndexPage.
