+++ jeffrey-pages/src/views/docs/hub/HubArchitectureOverviewPage.vue
@@ -164,7 +164,7 @@ onMounted(() => {
         prepares the Async-Profiler agent's JVM arguments before the application starts, and the
         <router-link to="/docs/agent/overview">Jeffrey Agent</router-link> writes JFR recordings
         to the shared filesystem the Server reads from. The
-        <router-link to="/docs/jib/overview">Jeffrey JIB</router-link> module packages the provisioner
+        <router-link to="/docs/jib">Jeffrey JIB</router-link> module packages the provisioner
         and Agent alongside your application image.
       </p>
